Remove version change model
Move only user code to tools
